Extract message send nodes into separate files
