+++ src/trufflesom/interpreter/nodes/MessageSendNode.java
@@ -1,20 +1,12 @@
 package trufflesom.interpreter.nodes;
 
-import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.DirectCallNode;
-import com.oracle.truffle.api.nodes.ExplodeLoop;
-import com.oracle.truffle.api.nodes.NodeCost;
 import com.oracle.truffle.api.source.SourceSection;
 
 import bd.primitives.Specializer;
 import bd.primitives.nodes.PreevaluatedExpression;
-import bd.tools.nodes.Invocation;
-import trufflesom.interpreter.TruffleCompiler;
-import trufflesom.interpreter.nodes.dispatch.AbstractDispatchNode;
-import trufflesom.interpreter.nodes.dispatch.DispatchChain.Cost;
-import trufflesom.interpreter.nodes.dispatch.GenericDispatchNode;
 import trufflesom.interpreter.nodes.dispatch.UninitializedDispatchNode;
 import trufflesom.primitives.Primitives;
 import trufflesom.vm.NotYetImplementedException;
@@ -72,150 +64,6 @@ public final class MessageSendNode {
         method.getCallTarget());
 
     return new SuperSendNode(selector, arguments, superMethodNode).initialize(source);
-  }
-
-  public abstract static class AbstractMessageSendNode extends ExpressionNode
-      implements PreevaluatedExpression, Invocation<SSymbol> {
-
-    @Children protected final ExpressionNode[] argumentNodes;
-
-    protected AbstractMessageSendNode(final ExpressionNode[] arguments) {
-      this.argumentNodes = arguments;
-    }
-
-    @Override
-    public final Object executeGeneric(final VirtualFrame frame) {
-      Object[] arguments = evaluateArguments(frame);
-      return doPreEvaluated(frame, arguments);
-    }
-
-    @ExplodeLoop
-    private Object[] evaluateArguments(final VirtualFrame frame) {
-      Object[] arguments = new Object[argumentNodes.length];
-      for (int i = 0; i < argumentNodes.length; i++) {
-        arguments[i] = argumentNodes[i].executeGeneric(frame);
-        assert arguments[i] != null;
-      }
-      return arguments;
-    }
-
-    public abstract int getNumberOfArguments();
-  }
-
-  public static final class UninitializedMessageSendNode extends AbstractMessageSendNode {
-
-    protected final SSymbol  selector;
-    protected final Universe universe;
-
-    protected UninitializedMessageSendNode(final SSymbol selector,
-        final ExpressionNode[] arguments, final Universe universe) {
-      super(arguments);
-      this.selector = selector;
-      this.universe = universe;
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName() + "(" + selector.getString() + ")";
-    }
-
-    @Override
-    public Object doPreEvaluated(final VirtualFrame frame, final Object[] arguments) {
-      return specialize(arguments).doPreEvaluated(frame, arguments);
-    }
-
-    private PreevaluatedExpression specialize(final Object[] arguments) {
-      TruffleCompiler.transferToInterpreterAndInvalidate("Specialize Message Node");
-
-      // We treat super sends separately for simplicity, might not be the
-      // optimal solution, especially in cases were the knowledge of the
-      // receiver class also allows us to do more specific things, but for the
-      // moment we will leave it at this.
-      // TODO: revisit, and also do more specific optimizations for super sends.
-
-      Primitives prims = universe.getPrimitives();
-
-      Specializer<Universe, ExpressionNode, SSymbol> specializer =
-          prims.getEagerSpecializer(selector, arguments, argumentNodes);
-
-      if (specializer != null) {
-        PreevaluatedExpression newNode =
-            (PreevaluatedExpression) specializer.create(arguments, argumentNodes,
-                sourceSection, universe);
-
-        return (PreevaluatedExpression) replace((ExpressionNode) newNode);
-      }
-
-      return makeGenericSend();
-    }
-
-    private GenericMessageSendNode makeGenericSend() {
-      GenericMessageSendNode send = new GenericMessageSendNode(selector, argumentNodes,
-          new UninitializedDispatchNode(selector, universe)).initialize(sourceSection);
-      return replace(send);
-    }
-
-    @Override
-    public SSymbol getInvocationIdentifier() {
-      return selector;
-    }
-
-    @Override
-    public int getNumberOfArguments() {
-      return selector.getNumberOfSignatureArguments();
-    }
-  }
-
-  // TODO: currently, we do not only specialize the given stuff above, but also what has been
-  // classified as 'value' sends in the OMOP branch. Is that a problem?
-
-  public static final class GenericMessageSendNode
-      extends AbstractMessageSendNode {
-
-    private final SSymbol selector;
-    private final int     numberOfSignatureArguments;
-
-    @Child private AbstractDispatchNode dispatchNode;
-
-    private GenericMessageSendNode(final SSymbol selector, final ExpressionNode[] arguments,
-        final AbstractDispatchNode dispatchNode) {
-      super(arguments);
-      this.selector = selector;
-      this.dispatchNode = dispatchNode;
-      this.numberOfSignatureArguments = selector.getNumberOfSignatureArguments();
-    }
-
-    @Override
-    public Object doPreEvaluated(final VirtualFrame frame,
-        final Object[] arguments) {
-      return dispatchNode.executeDispatch(frame, arguments);
-    }
-
-    public void replaceDispatchListHead(
-        final GenericDispatchNode replacement) {
-      CompilerAsserts.neverPartOfCompilation();
-      dispatchNode.replace(replacement);
-    }
-
-    @Override
-    public String toString() {
-      return "GMsgSend(" + selector.getString() + ")";
-    }
-
-    @Override
-    public NodeCost getCost() {
-      return Cost.getCost(dispatchNode);
-    }
-
-    @Override
-    public SSymbol getInvocationIdentifier() {
-      return selector;
-    }
-
-    @Override
-    public int getNumberOfArguments() {
-      return numberOfSignatureArguments;
-    }
   }
 
   public static final class SuperSendNode extends AbstractMessageSendNode {
