+++ src/trufflesom/interpreter/nodes/nary/UnaryExpressionNode.java
@@ -7,8 +7,8 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import trufflesom.interpreter.SomLanguage;
 import trufflesom.interpreter.bc.RespecializeException;
 import trufflesom.interpreter.nodes.ExpressionNode;
+import trufflesom.interpreter.nodes.GenericMessageSendNode;
 import trufflesom.interpreter.nodes.MessageSendNode;
-import trufflesom.interpreter.nodes.MessageSendNode.GenericMessageSendNode;
 import trufflesom.interpreter.nodes.bc.BytecodeLoopNode;
 import trufflesom.vm.VmSettings;
 import trufflesom.vmobjects.SSymbol;
